cli/requirements.txt
python-fire
prompt_toolkit
docstring_parser